Relocate all methods in Tasks into the respective game mode classes. Methods shared by several classes (like quit_task) are put in the base class Mode, or in specialized base classes like CameraMode (which houses zoom_camera, for example, which is used by ViewMode and ShotMode)
