+++ pooltool/ani/modes/view.py
@@ -3,14 +3,15 @@
 from pooltool.ani.modes import *
 
 
-class ViewMode(Mode):
-    keymap ={
+class ViewMode(CameraMode):
+    keymap = {
         action.aim: False,
         action.fine_control: False,
         action.move: True,
         action.quit: False,
         action.zoom: False,
     }
+
 
     def enter(self):
         self.mouse.hide()
@@ -31,5 +32,21 @@ class ViewMode(Mode):
     def exit(self):
         self.remove_task('view_task')
         self.remove_task('quit_task')
+
+
+    def view_task(self, task):
+        if self.keymap[action.aim]:
+            enter_kwargs = dict(
+                load_prev_cam = True
+            )
+            self.change_mode('aim', enter_kwargs=enter_kwargs)
+        elif self.keymap[action.zoom]:
+            self.zoom_camera()
+        elif self.keymap[action.move]:
+            self.move_camera()
+        else:
+            self.rotate_camera(cue_stick_too=False)
+
+        return task.cont
